Bugfix on extraction script to make sure some fields were properly exported in the CSV format. Also made cache response latencies equal to hit latencies. Added debug statement to figure out why L2 caches are having correct number of faulty blocks after CPUs are switched...

+++ src/mem/cache/tags/dpcs_lru.hh
@@ -271,7 +271,7 @@ public:
      */
     template <typename V>
     void forEachBlk(V &visitor) {
-        for (unsigned i = 0; i < numSets * assoc; ++i) {
+        for (unsigned long i = 0; i < numSets * assoc; ++i) {
             if (!visitor(blks[i]))
                 return;
         }
